+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/custom.py
@@ -26,10 +26,10 @@ from azure.graphrbac.models import GraphErrorException
 
 from azure.cli.core.util import get_file_json, shell_safe_json_parse
 
-from azure.graphrbac.models import (ApplicationCreateParameters, ApplicationUpdateParameters, PasswordCredential,
-                                    KeyCredential, UserCreateParameters, PasswordProfile,
-                                    ServicePrincipalCreateParameters, RequiredResourceAccess, AppRole,
-                                    ResourceAccess, GroupCreateParameters, CheckGroupMembershipParameters)
+from azure.graphrbac.models import (ApplicationCreateParameters, ApplicationUpdateParameters, AppRole,
+                                    PasswordCredential, KeyCredential, UserCreateParameters, PasswordProfile,
+                                    ServicePrincipalCreateParameters, RequiredResourceAccess, ResourceAccess,
+                                    GroupCreateParameters, CheckGroupMembershipParameters)
 
 from ._client_factory import _auth_client_factory, _graph_client_factory
 from ._multi_api_adaptor import MultiAPIAdaptor
@@ -668,7 +668,7 @@ def _resolve_group(client, identifier):
     return identifier
 
 
-def create_application(cmd, display_name, homepage=None, identifier_uris=None,
+def create_application(cmd, display_name, homepage=None, identifier_uris=None,  # pylint: disable=too-many-locals
                        available_to_other_tenants=False, password=None, reply_urls=None,
                        key_value=None, key_type=None, key_usage=None, start_date=None, end_date=None,
                        oauth2_allow_implicit_flow=None, required_resource_accesses=None, native_app=None,
@@ -691,19 +691,19 @@ def create_application(cmd, display_name, homepage=None, identifier_uris=None,
     if app_roles:
         app_roles = _build_app_roles(app_roles)
 
-    app_patch_param = ApplicationCreateParameters(available_to_other_tenants=available_to_other_tenants,
-                                                  display_name=display_name,
-                                                  identifier_uris=identifier_uris,
-                                                  homepage=homepage,
-                                                  reply_urls=reply_urls,
-                                                  key_credentials=key_creds,
-                                                  password_credentials=password_creds,
-                                                  oauth2_allow_implicit_flow=oauth2_allow_implicit_flow,
-                                                  required_resource_access=required_accesses,
-                                                  app_roles=app_roles)
+    app_create_param = ApplicationCreateParameters(available_to_other_tenants=available_to_other_tenants,
+                                                   display_name=display_name,
+                                                   identifier_uris=identifier_uris,
+                                                   homepage=homepage,
+                                                   reply_urls=reply_urls,
+                                                   key_credentials=key_creds,
+                                                   password_credentials=password_creds,
+                                                   oauth2_allow_implicit_flow=oauth2_allow_implicit_flow,
+                                                   required_resource_access=required_accesses,
+                                                   app_roles=app_roles)
 
     try:
-        result = graph_client.applications.create(app_patch_param)
+        result = graph_client.applications.create(app_create_param)
     except GraphErrorException as ex:
         if 'insufficient privileges' in str(ex).lower():
             link = 'https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-create-service-principal-portal'  # pylint: disable=line-too-long
@@ -715,14 +715,23 @@ def create_application(cmd, display_name, homepage=None, identifier_uris=None,
         # AAD graph doesn't have the API to create a native app, aka public client, the recommended hack is
         # to create a web app first, then convert to a native one
         # pylint: disable=protected-access
-        if 'public_client' not in ApplicationUpdateParameters._attribute_map:
-            ApplicationUpdateParameters._attribute_map['public_client'] = {'key': 'publicClient', 'type': 'bool'}
         app_patch_param = ApplicationUpdateParameters(identifier_uris=[])
-        setattr(app_patch_param, 'public_client', True)
+        app_patch_param.public_client = True
         graph_client.applications.patch(result.object_id, app_patch_param)
         result = graph_client.applications.get(result.object_id)
 
     return result
+
+
+def _get_grant_permissions(graph_client, client_sp_object_id=None, query_filter=None):
+    query_filter = query_filter or ("clientId eq '{}'".format(client_sp_object_id) if client_sp_object_id else None)
+    try:
+        grant_info = graph_client.oauth2_permission_grant.list(filter=query_filter)
+    except CloudError as ex:  # Graph doesn't follow the ARM error; otherwise would be caught by msrest-azure
+        if ex.status_code == 404:
+            return []
+        raise
+    return grant_info
 
 
 def list_permissions(cmd, identifier):
@@ -733,9 +742,7 @@ def list_permissions(cmd, identifier):
 
     # first get the permission grant history
     client_sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
-    grant_info = graph_client.oauth2.get(
-        filter="clientId eq '{}'".format(client_sp_object_id))  # pylint: disable=no-member
-    grant_histories = grant_info.additional_properties['value']
+    grant_permissions = _get_grant_permissions(graph_client, client_sp_object_id=client_sp_object_id)
 
     # get original permissions required by the application, we will cross check the history
     # and mark out granted ones
@@ -745,12 +752,28 @@ def list_permissions(cmd, identifier):
     for p in permissions:
         result = list(graph_client.service_principals.list(
             filter="servicePrincipalNames/any(c:c eq '{}')".format(p.resource_app_id)))
-        granted_times = 'N/A'
+        expiry_time = 'N/A'
         if result:
-            granted_times = ', '.join([x['startTime'] for x in grant_histories if
-                                       x['resourceId'] == result[0].object_id])
-        setattr(p, 'grantedTime', granted_times)
+            expiry_time = ', '.join([x.expiry_time for x in grant_permissions if
+                                     x.resource_id == result[0].object_id])
+        setattr(p, 'expiryTime', expiry_time)
     return permissions
+
+
+def list_permission_grants(cmd, identifier=None, query_filter=None, show_resource_name=None):
+    if identifier and query_filter:
+        raise CLIError('Please only use one of "--identifier" and "--filter", not both')
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    client_sp_object_id = None
+    if identifier:
+        client_sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
+    result = _get_grant_permissions(graph_client, client_sp_object_id=client_sp_object_id, query_filter=query_filter)
+    result = list(result)
+    if show_resource_name:
+        for r in result:
+            sp = graph_client.service_principals.get(r.resource_id)
+            setattr(r, 'resource_display_name', sp.display_name)
+    return result
 
 
 def add_permission(cmd, identifier, api, api_permissions):
@@ -761,9 +784,14 @@ def add_permission(cmd, identifier, api, api_permissions):
     for e in api_permissions:
         access_id, access_type = e.split('=')
         resource_accesses.append(ResourceAccess(id=access_id, type=access_type))
-    required_resource_access = RequiredResourceAccess(resource_app_id=api,
-                                                      resource_access=resource_accesses)
-    existing.append(required_resource_access)
+
+    existing_resource_access = next((e for e in existing if e.resource_app_id == api), None)
+    if existing_resource_access:
+        existing_resource_access.resource_access += resource_accesses
+    else:
+        required_resource_access = RequiredResourceAccess(resource_app_id=api,
+                                                          resource_access=resource_accesses)
+        existing.append(required_resource_access)
     update_parameter = ApplicationUpdateParameters(required_resource_access=existing)
     graph_client.applications.patch(application.object_id, update_parameter)
     logger.warning('Invoking "az ad app permission grant --id %s --api %s" is needed to make the '
@@ -779,7 +807,8 @@ def delete_permission(cmd, identifier, api):
     return graph_client.applications.patch(application.object_id, update_parameter)
 
 
-def grant_application(cmd, identifier, api, expires='1', scope='user_impersonation'):
+def grant_application(cmd, identifier, api, consent_type=None, principal_id=None,
+                      expires='1', scope='user_impersonation'):
     graph_client = _graph_client_factory(cmd.cli_ctx)
 
     # Get the Service Principal ObjectId for the client app
@@ -787,6 +816,13 @@ def grant_application(cmd, identifier, api, expires='1', scope='user_impersonati
 
     # Get the Service Principal ObjectId for associated app
     associated_sp_object_id = _resolve_service_principal(graph_client.service_principals, api)
+
+    # ensure to remove the older grant
+    grant_permissions = _get_grant_permissions(graph_client, client_sp_object_id=client_sp_object_id)
+    to_delete = [p.object_id for p in grant_permissions if p.client_id == client_sp_object_id and
+                 p.resource_id == associated_sp_object_id]
+    for p in to_delete:
+        graph_client.oauth2_permission_grant.delete(p)
 
     # Build payload
     start_date = datetime.datetime.utcnow()
@@ -803,24 +839,22 @@ def grant_application(cmd, identifier, api, expires='1', scope='user_impersonati
     payload = {
         "odata.type": "Microsoft.DirectoryServices.OAuth2PermissionGrant",
         "clientId": client_sp_object_id,
-        "consentType": "AllPrincipals",
+        "consentType": consent_type,
         "resourceId": associated_sp_object_id,
         "scope": scope,
+        'principalId': principal_id,
         "startTime": start_date.isoformat(),
         "expiryTime": end_date.isoformat()
     }
 
     # Grant OAuth2 permissions
-    response = graph_client.oauth2.grant(payload)  # pylint: disable=no-member
-
-    return response
+    return graph_client.oauth2_permission_grant.create(payload)  # pylint: disable=no-member
 
 
 def update_application(instance, display_name=None, homepage=None,  # pylint: disable=unused-argument
                        identifier_uris=None, password=None, reply_urls=None, key_value=None,
                        key_type=None, key_usage=None, start_date=None, end_date=None, available_to_other_tenants=None,
                        oauth2_allow_implicit_flow=None, required_resource_accesses=None, app_roles=None):
-    from azure.cli.core.commands.arm import make_camel_case, make_snake_case
     password_creds, key_creds, required_accesses = None, None, None
     if any([password, key_value]):
         password_creds, key_creds = _build_application_creds(password, key_value, key_type,
@@ -832,23 +866,16 @@ def update_application(instance, display_name=None, homepage=None,  # pylint: di
     if app_roles:
         app_roles = _build_app_roles(app_roles)
 
-    # Workaround until https://github.com/Azure/azure-rest-api-specs/issues/3437 is fixed
-    def _get_property(name):
-        try:
-            return getattr(instance, make_snake_case(name))
-        except AttributeError:
-            return instance.additional_properties.get(make_camel_case(name), None)
-
     app_patch_param = ApplicationUpdateParameters(
-        display_name=display_name or _get_property('display_name'),
-        homepage=homepage or _get_property('homepage'),
-        identifier_uris=identifier_uris or _get_property('identifier_uris'),
-        reply_urls=reply_urls or _get_property('reply_urls'),
-        key_credentials=key_creds or None,
-        password_credentials=password_creds or None,
-        available_to_other_tenants=available_to_other_tenants or _get_property('available_to_other_tenants'),
-        required_resource_access=required_accesses or _get_property('required_resource_access'),
-        oauth2_allow_implicit_flow=oauth2_allow_implicit_flow or _get_property('oauth2_allow_implicit_flow'),
+        display_name=display_name,
+        homepage=homepage,
+        identifier_uris=identifier_uris,
+        reply_urls=reply_urls,
+        key_credentials=key_creds,
+        password_credentials=password_creds,
+        available_to_other_tenants=available_to_other_tenants,
+        required_resource_access=required_accesses,
+        oauth2_allow_implicit_flow=oauth2_allow_implicit_flow,
         app_roles=app_roles)
 
     return app_patch_param
@@ -858,6 +885,12 @@ def patch_application(cmd, identifier, parameters):
     graph_client = _graph_client_factory(cmd.cli_ctx)
     object_id = _resolve_application(graph_client.applications, identifier)
     return graph_client.applications.patch(object_id, parameters)
+
+
+def patch_service_principal(cmd, identifier, parameters):
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    object_id = _resolve_service_principal(graph_client.service_principals, identifier)
+    return graph_client.service_principals.update(object_id, parameters)
 
 
 def _build_application_accesses(required_resource_accesses):
@@ -1467,9 +1500,9 @@ def reset_service_principal_credential(cmd, name, password=None, create_cert=Fal
             custom_key_identifier=custom_key_identifier
         ))
 
-    app_create_param = ApplicationUpdateParameters(password_credentials=app_creds, key_credentials=cert_creds)
+    app_patch_param = ApplicationUpdateParameters(password_credentials=app_creds, key_credentials=cert_creds)
 
-    client.applications.patch(app.object_id, app_create_param)
+    client.applications.patch(app.object_id, app_patch_param)
 
     result = {
         'appId': app.app_id,