+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,8 @@
 
 Release History
 ===============
+* `ad app permission add`: support to add permissions on existing api
+* `ad app permission list`: fix a bad error when there is no permissions
 * `ad sp delete`: skip role assignment delete if the current account has no subscription
 
 2.4.2